requirements.txt
Django
python-telegram-bot
polib
gunicorn
psycopg2-binary
python-dotenv